split monolith two authentication microservice and ordering microservice
