+++ authentication/src/main/kotlin/com/ali/authentication/AuthenticationApplication.kt
@@ -1,0 +1,11 @@
+package com.ali.authentication
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class AuthenticationApplication
+
+fun main(args: Array<String>) {
+    runApplication<AuthenticationApplication>(*args)
+}
