+++ authentication/src/main/kotlin/com/ali/authentication/CustomUserDetailsService.kt
@@ -1,0 +1,21 @@
+package com.ali.authentication
+
+
+import com.ali.authentication.user.UserRepository
+import org.springframework.security.core.userdetails.*
+import org.springframework.stereotype.Service
+
+@Service
+class CustomUserDetailsService(
+    private val usersRepository: UserRepository
+) : UserDetailsService {
+    override fun loadUserByUsername(username: String): UserDetails {
+        val user = usersRepository.findByUsername(username)
+            ?: throw UsernameNotFoundException("User not found")
+
+        return User.builder()
+            .username(user.username)
+            .password(user.password)
+            .build()
+    }
+}
