+++ authentication/src/main/kotlin/com/ali/authentication/profile/ProfileEntity.kt
@@ -1,0 +1,21 @@
+package com.ali.authentication.profile
+
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+
+@Entity
+@Table(name = "profiles")
+data class ProfileEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long? = null,
+
+    var firstName: String,
+    var lastName: String,
+    var phoneNumber: String,
+
+    var userId: Long,
+)
