+++ authentication/src/main/kotlin/com/ali/authentication/profile/ProfileRepository.kt
@@ -1,0 +1,9 @@
+package com.ali.authentication.profile
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ProfileRepository : JpaRepository<ProfileEntity, Long> {
+    fun findByUserId(userId: Long): ProfileEntity?
+}
