+++ authentication/src/main/kotlin/com/ali/authentication/user/UserEntity.kt
@@ -1,0 +1,20 @@
+package com.ali.authentication.user
+
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
+
+@Entity
+@Table(name = "users")
+data class UserEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long? = null,
+
+    var name: String,
+    var email: String,
+    val username: String,
+    val password: String,
+)
