+++ authentication/src/main/kotlin/com/ali/authentication/user/UserRepository.kt
@@ -1,0 +1,10 @@
+package com.ali.authentication.user
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UserRepository : JpaRepository<UserEntity, Long> {
+    fun findByUsername(username: String) : UserEntity?
+
+}
