+++ order/src/main/kotlin/coded/spring/OrderingApplication.kt
@@ -1,0 +1,14 @@
+package coded.spring
+
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+
+@SpringBootApplication
+class OrderingApplication
+
+fun main(args: Array<String>) {
+    runApplication<OrderingApplication>(*args)
+}
+
