+++ order/src/main/kotlin/coded/spring/entity/MenuEntity.kt
@@ -1,0 +1,16 @@
+package coded.spring.entity
+
+import jakarta.persistence.*
+import java.math.BigDecimal
+
+@Entity
+@Table(name = "menu")
+data class MenuEntity(
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long? = null,
+
+    var name: String,
+    var price: BigDecimal
+)
