+++ order/src/main/kotlin/coded/spring/entity/OrderEntity.kt
@@ -1,0 +1,15 @@
+package coded.spring.entity
+
+import jakarta.persistence.*
+
+@Entity
+@Table(name = "orders")
+data class OrderEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long? = null,
+
+    var restaurant: String,
+
+    var userId: Long
+)
