+++ order/src/main/kotlin/coded/spring/repository/ItemRepository.kt
@@ -1,0 +1,10 @@
+package coded.spring.repository
+
+import coded.spring.entity.ItemEntity
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface ItemRepository : JpaRepository<ItemEntity, Long>{
+
+}
