+++ order/src/main/kotlin/coded/spring/repository/MenuRepository.kt
@@ -1,0 +1,10 @@
+package coded.spring.repository
+
+import coded.spring.entity.MenuEntity
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface MenuRepository : JpaRepository<MenuEntity, Long>{
+
+}
