+++ order/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+spring.application.name=Kotlin.SpringbootV2
+
+server.port = 4444
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/OrderingDB
+spring.datasource.username=postgres
+spring.datasource.password=alix
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.show-sql=true
+spring.jpa.properties.hibe1rnate.format_sql=true
+
+springdoc.api-docs.path=/api-docs
