--- src/main/kotlin/com/coded/spring/authentication/CustomUserDetailsService.kt
@@ -1,21 +1,0 @@
-package com.coded.spring.authentication
-
-
-import com.coded.spring.repository.UserRepository
-import org.springframework.security.core.userdetails.*
-import org.springframework.stereotype.Service
-
-@Service
-class CustomUserDetailsService(
-    private val usersRepository: UserRepository
-) : UserDetailsService {
-    override fun loadUserByUsername(username: String): UserDetails {
-        val user = usersRepository.findByUsername(username)
-            ?: throw UsernameNotFoundException("User not found")
-
-        return User.builder()
-            .username(user.username)
-            .password(user.password)
-            .build()
-    }
-}
