--- src/main/kotlin/com/coded/spring/authentication/jwt/JwtAuthenticationFilter.kt
@@ -1,46 +1,0 @@
-package com.coded.spring.authentication.jwt
-
-
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.*
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.security.core.userdetails.UserDetailsService
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-
-@Component
-class JwtAuthenticationFilter(
-    private val jwtService: JwtService,
-    private val userDetailsService: UserDetailsService
-) : OncePerRequestFilter() {
-
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain
-    ) {
-        val authHeader = request.getHeader("Authorization")
-        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-            filterChain.doFilter(request, response)
-            return
-        }
-
-        val token = authHeader.substring(7)
-        val username = jwtService.extractUsername(token)
-
-        if (SecurityContextHolder.getContext().authentication == null) {
-            if (jwtService.isTokenValid(token, username)) {
-                val userDetails = userDetailsService.loadUserByUsername(username)
-                val authToken = UsernamePasswordAuthenticationToken(
-                    userDetails, null, userDetails.authorities
-                )
-                authToken.details = WebAuthenticationDetailsSource().buildDetails(request)
-                SecurityContextHolder.getContext().authentication = authToken
-            }
-        }
-
-        filterChain.doFilter(request, response)
-    }
-}
