--- src/main/kotlin/com/coded/spring/entity/UserEntity.kt
@@ -1,22 +1,0 @@
-package com.coded.spring.entity
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-import jakarta.persistence.OneToMany
-import jakarta.persistence.Table
-
-@Entity
-@Table(name = "users")
-data class UserEntity(
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null,
-
-    var name: String,
-    var email: String,
-    val username: String,
-    val password: String,
-)
