--- src/main/kotlin/com/coded/spring/repository/ItemRepository.kt
@@ -1,10 +1,0 @@
-package com.coded.spring.repository
-
-import com.coded.spring.entity.ItemEntity
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ItemRepository : JpaRepository<ItemEntity, Long>{
-
-}
