--- src/main/kotlin/com/coded/spring/repository/MenuRepository.kt
@@ -1,10 +1,0 @@
-package com.coded.spring.repository
-
-import com.coded.spring.entity.MenuEntity
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface MenuRepository : JpaRepository<MenuEntity, Long>{
-
-}
