--- src/main/kotlin/com/coded/spring/repository/ProfileRepository.kt
@@ -1,10 +1,0 @@
-package com.coded.spring.repository
-
-import com.coded.spring.entity.ProfileEntity
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface ProfileRepository : JpaRepository<ProfileEntity, Long> {
-    fun findByUserId(userId: Long): ProfileEntity?
-}
