--- src/main/kotlin/com/coded/spring/repository/UserRepository.kt
@@ -1,12 +1,0 @@
-package com.coded.spring.repository
-
-
-import com.coded.spring.entity.UserEntity
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface UserRepository : JpaRepository<UserEntity, Long> {
-    fun findByUsername(username: String) : UserEntity?
-
-}
